[207] Tools gefixt. Jetzt standardmäßig Pickaxe im Inventar und in der Tool-Leiste

+++ OctoAwesome/OctoAwesome.Client/Components/HudComponent.cs
@@ -98,8 +98,8 @@ namespace OctoAwesome.Client.Components
             get
             {
                 return new Index2(
-                  (int)GraphicsDevice.Viewport.Width,
-                  (int)GraphicsDevice.Viewport.Height);
+                  GraphicsDevice.Viewport.Width,
+                  GraphicsDevice.Viewport.Height);
             }
         }
